Revert "Use uint64_t for branch weights instead of uint32_t"
This reverts commit 10f2a0d662d8d72eaac48d3e9b31ca8dc90df5a4.

More uint64_t overflows.

+++ compiler-rt/test/profile/instrprof-dump.c
@@ -58,5 +58,5 @@ __attribute__((noinline)) int bar(int X) {
 
 /*
 PROF: ![[ENT]] = !{!"function_entry_count", i64 2}  
-PROF: ![[PD1]] = !{!"branch_weights", i64 2, i64 2}
+PROF: ![[PD1]] = !{!"branch_weights", i32 2, i32 2}
 */
